Stop passing authentication params as normal parameters
OpenAPI has support for declaring what, if any, authentication
mechanisms that need to be in place for API endpoints. Use this
mechanism so one can configure the generated API client once, instead
of passing in the authorization and subscription key on each and every
call.  This leads to much more ergonomic clients.

Also stop passing the content-type as a normal parameter, since this
is handled correctly by client generators already.